Refactor DockAppToggler's main.swift for improved UI and memory management
- Removed extensive constant definitions and logging functionality to streamline the codebase.
- Simplified the WindowInfo struct and enhanced button creation logic in WindowChooserView for better UI clarity.
- Updated CloseButton and MinimizeButton classes to improve performance by removing unnecessary memory tracking.
- Enhanced hover effects and button responsiveness, contributing to a more intuitive user experience.

These changes prepare the application for future enhancements while improving overall performance and maintainability.
