+++ Sources/DockAppToggler/UI/StatusBarController.swift
@@ -1,0 +1,108 @@
+import AppKit
+import Sparkle
+import Cocoa
+
+@MainActor
+class StatusBarController {
+    private var statusBar: NSStatusBar
+    private var statusItem: NSStatusItem
+    private var menu: NSMenu
+    private weak var updaterController: SPUStandardUpdaterController?
+    private let autostartMenuItem: NSMenuItem
+    
+    init(updater: SPUStandardUpdaterController?) {
+        statusBar = NSStatusBar.system
+        statusItem = statusBar.statusItem(withLength: NSStatusItem.squareLength)
+        menu = NSMenu()
+        
+        // Store the shared updater controller
+        updaterController = updater
+        
+        // Create autostart menu item
+        autostartMenuItem = NSMenuItem(
+            title: "Start at Login",
+            action: #selector(toggleAutostart),
+            keyEquivalent: ""
+        )
+        
+        if let button = statusItem.button {
+            // Try multiple paths to find the icon
+            let iconImage: NSImage?
+            if let bundleIconPath = Bundle.main.path(forResource: "trayicon", ofType: "png") {
+                // App bundle path
+                iconImage = NSImage(contentsOfFile: bundleIconPath)
+            } else {
+                // Development path
+                let devIconPath = "Sources/DockAppToggler/Resources/trayicon.png"
+                iconImage = NSImage(contentsOfFile: devIconPath)
+            }
+            
+            if let image = iconImage {
+                // Create a copy of the image at the desired size
+                let resizedImage = NSImage(size: NSSize(width: 18, height: 18))
+                resizedImage.lockFocus()
+                image.draw(in: NSRect(origin: .zero, size: NSSize(width: 18, height: 18)))
+                resizedImage.unlockFocus()
+                
+                // Set as template
+                resizedImage.isTemplate = true
+                button.image = resizedImage
+            }
+        }
+        
+        setupMenu()
+        updateAutostartState()
+    }
+    
+    private func setupMenu() {
+        // Add autostart toggle
+        autostartMenuItem.target = self
+        menu.addItem(autostartMenuItem)
+        
+        // Add separator
+        menu.addItem(NSMenuItem.separator())
+        
+        // Add update menu item only if updater is available
+        if let updaterController = updaterController {
+            let updateItem = NSMenuItem(title: "Check for Updates...", 
+                                      action: #selector(SPUStandardUpdaterController.checkForUpdates(_:)), 
+                                      keyEquivalent: "")
+            updateItem.target = updaterController
+            menu.addItem(updateItem)
+            menu.addItem(NSMenuItem.separator())
+        }
+        
+        // Add restart option
+        let restartItem = NSMenuItem(
+            title: "Restart",
+            action: #selector(restartApp),
+            keyEquivalent: "r"
+        )
+        restartItem.target = self
+        menu.addItem(restartItem)
+        
+        menu.addItem(NSMenuItem.separator())
+        
+        let quitItem = NSMenuItem(title: "Quit", 
+                                action: #selector(NSApplication.terminate(_:)), 
+                                keyEquivalent: "q")
+        quitItem.target = NSApp
+        menu.addItem(quitItem)
+        
+        statusItem.menu = menu
+    }
+    
+    @objc private func toggleAutostart() {
+        let isEnabled = !LoginItemManager.shared.isLoginItemEnabled
+        LoginItemManager.shared.setLoginItemEnabled(isEnabled)
+        updateAutostartState()
+    }
+    
+    private func updateAutostartState() {
+        autostartMenuItem.state = LoginItemManager.shared.isLoginItemEnabled ? .on : .off
+    }
+    
+    @objc private func restartApp() {
+        NSApplication.restart()
+    }
+}
